38732
restructured files a bit so it is easier to follow

added collection,project data components
