+++ packages/dina-ui/pages/managed-attribute/list.tsx
@@ -18,7 +18,7 @@ import PageLayout from "../../components/page/PageLayout";
 import {
   COLLECTION_MODULE_TYPE_LABELS,
   CollectionModuleType,
-  MANAGED_ATTRIBUTE_TYPE_OPTIONS,
+  VOCABULARY_ELEMENT_TYPE_OPTIONS,
   ManagedAttribute
 } from "../../types/collection-api";
 
@@ -199,7 +199,7 @@ function GenericManagedAttributeListView({
         const labelKey: keyof typeof DINAUI_MESSAGES_ENGLISH | undefined =
           acceptedValues?.length
             ? "field_vocabularyElementType_picklist_label"
-            : MANAGED_ATTRIBUTE_TYPE_OPTIONS.find(
+            : VOCABULARY_ELEMENT_TYPE_OPTIONS.find(
                 (option) => option.value === vocabularyElementType
               )?.labelKey;
