+++ packages/dina-ui/pages/seqdb/managed-attribute/edit.tsx
@@ -8,11 +8,13 @@ import {
 } from "../../../components";
 import { DinaMessage, useDinaIntl } from "../../../intl/dina-ui-intl";
 import {
-  ManagedAttribute,
+  ManagedAttribute
+} from "../../../types/collection-api";
+import {
   SEQDB_MODULE_TYPE_LABELS,
   SEQDB_MODULE_TYPES,
   SeqDBModuleType
-} from "../../../types/collection-api";
+} from "../../../types/seqdb-api";
 import PageLayout from "packages/dina-ui/components/page/PageLayout";
 
 export function ManagedAttributesEditPage({ router }: WithRouterProps) {
