+++ packages/dina-ui/types/collection-api/resources/ManagedAttribute.ts
@@ -1,7 +1,7 @@
 import { KitsuResource } from "kitsu";
-import { DINAUI_MESSAGES_ENGLISH } from "../../../intl/dina-ui-en";
 import { MultilingualDescription } from "../../common";
 import { HasDinaMetaInfo } from "../../DinaJsonMetaInfo";
+import { VocabularyElementType } from "./VocabularyElementType";
 
 export interface ManagedAttributeAttributes<TComponent = string> {
   type: "managed-attribute";
@@ -16,73 +16,6 @@ export interface ManagedAttributeAttributes<TComponent = string> {
   createdOn?: string;
   multilingualDescription?: MultilingualDescription;
 }
-
-export type VocabularyElementType =
-  | "INTEGER"
-  | "DECIMAL"
-  | "STRING"
-  | "PICKLIST"
-  | "DATE"
-  | "BOOL";
-
-export const COLLECTION_MODULE_TYPES = [
-  "ASSEMBLAGE",
-  "COLLECTING_EVENT",
-  "DETERMINATION",
-  "MATERIAL_SAMPLE",
-  "ORGANISM",
-  "PREPARATION",
-  "SITE"
-] as const;
-export type CollectionModuleType = (typeof COLLECTION_MODULE_TYPES)[number];
-export const COLLECTION_MODULE_TYPE_LABELS: Record<
-  CollectionModuleType,
-  string
-> = {
-  ASSEMBLAGE: "assemblage",
-  COLLECTING_EVENT: "collectingEvent",
-  DETERMINATION: "determination",
-  MATERIAL_SAMPLE: "materialSample",
-  ORGANISM: "organism",
-  PREPARATION: "preparation",
-  SITE: "site"
-};
-
-export const SEQDB_MODULE_TYPES = ["GENERIC_MOLECULAR_ANALYSIS"] as const;
-export type SeqDBModuleType = (typeof SEQDB_MODULE_TYPES)[number];
-export const SEQDB_MODULE_TYPE_LABELS: Record<SeqDBModuleType, string> = {
-  GENERIC_MOLECULAR_ANALYSIS: "genericMolecularAnalysis"
-};
-
-export const MANAGED_ATTRIBUTE_TYPE_OPTIONS: {
-  labelKey: keyof typeof DINAUI_MESSAGES_ENGLISH;
-  value: VocabularyElementType;
-}[] = [
-  {
-    labelKey: "field_vocabularyElementType_integer_label",
-    value: "INTEGER"
-  },
-  {
-    labelKey: "field_vocabularyElementType_decimal_label",
-    value: "DECIMAL"
-  },
-  {
-    labelKey: "field_vocabularyElementType_text_label",
-    value: "STRING"
-  },
-  {
-    labelKey: "field_vocabularyElementType_picklist_label",
-    value: "PICKLIST"
-  },
-  {
-    labelKey: "field_vocabularyElementType_date_label",
-    value: "DATE"
-  },
-  {
-    labelKey: "field_vocabularyElementType_boolean_label",
-    value: "BOOL"
-  }
-];
 
 export type ManagedAttribute<TComponent = string> = KitsuResource &
   ManagedAttributeAttributes<TComponent> &
