+++ packages/dina-ui/types/seqdb-api/index.ts
@@ -2,6 +2,7 @@ export * from "./resources/PcrBatch";
 export * from "./resources/PcrBatchItem";
 export * from "./resources/PcrPrimer";
 export * from "./resources/SeqBatch";
+export * from "./resources/SeqDBModuleType";
 export * from "./resources/ThermocyclerProfile";
 export * from "./resources/Product";
 export * from "./resources/Region";
